Revert "move RoutingContext tests to their own file"
didn't look closely enough, turns out these are RoutingContext tests!

This reverts commit 824c816c306062c1595d4ea6de2ff409c1c28897.
